+++ src/_pyorc/Writer.cpp
@@ -157,7 +157,7 @@ Writer::close()
 }
 
 void
-Writer::addMetadata(py::str key, py::bytes value)
+Writer::addUserMetadata(py::str key, py::bytes value)
 {
     writer->addUserMetadata(py::cast<std::string>(key), py::str(value));
 }