+++ src/pyorc/writer.py
@@ -36,7 +36,7 @@ class Writer(writer):
         if 0.0 >= bloom_filter_fpp or bloom_filter_fpp >= 1.0:
             raise ValueError("False positive probability should be > 0.0 & < 1.0")
         self.__schema = schema
-        self.__metadata = {}
+        self.__user_metadata = {}
         comp = CompressionKind(compression)
         comp_strat = CompressionStrategy(compression_strategy)
         bf_set = set()
@@ -78,21 +78,21 @@ class Writer(writer):
         self.close()
 
     def close(self) -> None:
-        for key, val in self.__metadata.items():
-            super()._add_metadata(key, val)
+        for key, val in self.__user_metadata.items():
+            super()._add_user_metadata(key, val)
         super().close()
 
     @property
     def schema(self) -> TypeDescription:
         return copy.deepcopy(self.__schema)
 
-    def set_metadata(self, **kwargs) -> None:
+    def set_user_metadata(self, **kwargs) -> None:
         for key, val in kwargs.items():
             if not isinstance(val, bytes):
                 raise TypeError(
                     "All values must be bytes, key '{0}' is {1}".format(key, type(val))
                 )
-            self.__metadata[key] = val
+            self.__user_metadata[key] = val
 
     def writerows(self, rows: Iterable) -> int:
         num = 0